fix delete request cors issues

+++ backend/src/Controllers/Controllers.php
@@ -9,8 +9,23 @@ use Scandiweb\Test\Controllers\Utils\HttpResponse;
 
 class Controllers 
 {
+  private static function isValidMethod(string $method): void
+  {
+    if (
+      isset($_SERVER['REQUEST_METHOD']) &&
+      in_array($_SERVER['REQUEST_METHOD'], Constants::ALLOWED_METHODS)
+      ) 
+    {
+      $actualMethod = $_SERVER['REQUEST_METHOD'];
+      $responseMsg = $actualMethod . ' not allowed on this route!!!';
+      strtolower($method) !== strtolower($actualMethod) &&
+      HttpResponse::notAllowed($responseMsg);
+    }
+  }
+
   public static function getProducts(): void
   {
+    self::isValidMethod('GET');
     try 
     {
       $dbObj = new DbConnect;
@@ -36,14 +51,15 @@ class Controllers
     }
     catch (\Exception $e)
     {
-      echo HttpResponse::set('db-error');
-      exit;
+      HttpResponse::dbError($e->getMessage());
     }
   }
 
 
   public static function addProduct(): void
   {
+    self::isValidMethod('POST');
+    
     $data = $_POST;
     $p = ['price', 'size', 'weight', 'height', 'width', 'length'];
     $keys = array_keys($data);
@@ -52,41 +68,40 @@ class Controllers
     }
     
     $class = "Scandiweb\\Test\\Models\\".ucfirst($data['type']);
-    try 
-    {
-      $product = new $class(...array_values($data));
-      echo $product->save();
-    }
-    catch (\ErrorException $e)
-    {
-      echo HttpResponse::set('invalid-data');
-    }
+    $product = new $class(...array_values($data));
+    $product->save();
   }
 
 
   public static function deleteProducts(): void
-  {    
+  { 
+    self::isValidMethod('DELETE');
+
+    $data = (array) json_decode(file_get_contents("php://input"));
+    $sqls = [];
+
+    foreach(array_keys($data) as $db) {
+      if (count($data[$db]) > 0) {
+        $skus = implode(',', array_map(fn($item) => "'$item'", $data[$db]));
+        $sql = "DELETE FROM $db WHERE sku IN ($skus)";
+        array_push($sqls, $sql);
+      }
+    }
     try 
     {
       $dbObj = new DbConnect;
       $dbConn = $dbObj->connect();
-      
-      $data = $_REQUEST;
 
-      foreach(array_keys($data) as $db) {
-        $skus = implode(
-          ',', 
-          array_map(fn($item) => "'$item'", $data[$db])
-        );
-        $sql = "DELETE FROM $db WHERE sku IN ($skus)";
+      foreach($sqls as $sql){
         $stmt = $dbConn->prepare($sql);
         $stmt->execute();
       }
-      echo HttpResponse::set('deleted');
+
+      HttpResponse::deleted();
     } 
     catch (\Exception $e) 
     {
-      echo $e->getMessage();
+      HttpResponse::dbError($e->getMessage());
     }
   }
 }
